prepare add role

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->string('picture');
-            $table->integer('identity')->nullable();
+            $table->bigInteger('identity')->nullable();
             $table->string('name');
             $table->string('email');
             $table->string('kelas')->nullable();
